Rename --title option to --site-title for clarity
It seems a common mistake would be to assume that "title" means the
title of the glog itself. I observed this in the wild when a user set up
a glog and what should be a glog title appeared in the site breadcrumb
link in the index. I think the option name "title" was misleading.

Now --site-title, not --title, is responsible for the text of the
breadcrumb link up from the index page. It has the same default value as
before.

--feed-title is responsible for the Atom feed title text, and defaults
to --site-title.

Finally, a new --title option lets you customize the index page heading
and post page breadcrumb links, which you probably thought was what
--title did in the first place. The default value is "Glog", which was
previously baked into the templates.